requirements.txt
aiogram<=2.25.2
requests
loguru
python-dotenv
discord.py

